Api

TermoQuimica.general.P_Antoine — Function

P_Antoine(CA, T, i)

\[\log_{10}{P(mmHg)}=A-\frac{B}{C+T(°C)}\]

Los campos de entrada son:

CA :: n×3 Array{Float,2} Constantes de Antoine para los n componentes

T :: Float Es la temperatura del sistema en °C

i :: Int Componente i al que se quiere conocer su presión

Salida:

Regresa la presión P en mmHg

P :: m-element Array{Float64,1} presión

TermoQuimica.general.T_Antoine — Function

T_Antoine(CA, P, i)

\[T(°C)=-C-\frac{B}{log_{10}{P(mmHg)}-A}\]

Los campos de entrada son:

CA :: n×3 Array{Float,2} Constantes de Antoine para los n componentes

P :: Float Es la presión del sistema en mmHg

i :: Int Componente i al que se quiere conocer su temperatura

Salida:

Regresa la temperatura T en °C

T :: m-element Array{Float64,1} temperatura

TermoQuimica.ideal.pxy — Function

pxy(n,T,CA,xx;uni=u"Torr")

Los campos de entrada son:

n :: Int Es el numero de componentes

T :: Float Es la temperatura del sistema

CA :: n×3 Array{Float,2} Constantes de Antoine para los n componentes

xx :: StepRange ó Vector Son los puntos en el liquido donde se buscara el equilibrio con el vapor

uni Son las unidades en las que se desea el resultado, por defecto esta en Torr

Salida:

Regresa las fracciones de liquido x, las fracciones del vapor y y la presión P en Torr (m es la longitud de xx).

x :: n×m Array{Float64,2} liquido

y :: n×m Array{Float64,2} vapor

P :: m-element Array{Float64,1} presión

TermoQuimica.ideal.txy — Function

txy(n,Pobj,CA,xx;uni=u"K")

Los campos de entrada son:

n :: Int Es el numero de componentes

Pobj :: Float Es la presión del sistema

CA :: n×3 Array{Float,2} Constantes de Antoine para los n componentes

xx :: StepRange ó Vector Son los puntos en el liquido donde se buscara el equilibrio con el vapor

uni Son las unidades en las que se desea el resultado, por defecto esta en grados kelvin

Salida:

Regresa las fracciones de liquido x, las fracciones del vapor y y la temperatura T en grados kelvin (m es la longitud de xx).

x :: n×m Array{Float64,2} liquido

y :: n×m Array{Float64,2} vapor

T :: m-element Array{Float64,1} temperatura

TermoQuimica.margules.pxy — Function

pxy(n,T,CA,Λ,xx;uni=u"Torr")

Los campos de entrada son:

n :: Int Es el numero de componentes

T :: Float Es la temperatura del sistema

CA :: n×3 Array{Float,2} Constantes de Antoine para los n componentes

Λ :: n×n Array{Float,2} Parámetros de interacción entre los componentes en grados kelvin

xx :: StepRange ó Vector Son los puntos en el liquido donde se buscara el equilibrio con el vapor

uni Son las unidades en las que se desea el resultado, por defecto esta en Torr

Salida:

Regresa las fracciones de liquido x, las fracciones del vapor y y la presión P en Torr (m es la longitud de xx).

x :: n×m Array{Float64,2} liquido

y :: n×m Array{Float64,2} vapor

P :: m-element Array{Float64,1} presión

TermoQuimica.margules.txy — Function

txy(n,Pobj,CA,Λ,xx;uni=u"K")

Los campos de entrada son:

n :: Int Es el numero de componentes

Pobj :: Float Es la presión del sistema

CA :: n×3 Array{Float,2} Constantes de Antoine para los n componentes

Λ :: n×n Array{Float,2} Parámetros de interacción entre los componentes en grados kelvin

xx :: StepRange ó Vector Son los puntos en el liquido donde se buscara el equilibrio con el vapor

uni Son las unidades en las que se desea el resultado, por defecto esta en grados kelvin

Salida:

Regresa las fracciones de liquido x, las fracciones del vapor y y la temperatura T en grados kelvin (m es la longitud de xx).

x :: n×m Array{Float64,2} liquido

y :: n×m Array{Float64,2} vapor

T :: m-element Array{Float64,1} temperatura

TermoQuimica.vanlaar.pxy — Function

pxy(n,T,CA,Λ,xx;uni=u"Torr")

Los campos de entrada son:

n :: Int Es el numero de componentes

T :: Float Es la temperatura del sistema

CA :: n×3 Array{Float,2} Constantes de Antoine para los n componentes

Λ :: n×n Array{Float,2} Parametros de interacción entre los componentes en grados kelvin

xx :: StepRange ó Vector Son los puntos en el liquido donde se buscara el equilibrio con el vapor

uni Son las unidades en las que se desea el resultado, por defecto esta en Torr

Salida:

Regresa las fracciones de liquido x, las fracciones del vapor y y la presión P en Torr (m es la longitud de xx).

x :: n×m Array{Float64,2} liquido

y :: n×m Array{Float64,2} vapor

P :: m-element Array{Float64,1} presión

TermoQuimica.vanlaar.txy — Function

txy(n,Pobj,CA,Λ,xx;uni=u"K")

Los campos de entrada son:

n :: Int Es el numero de componentes

Pobj :: Float Es la presión del sistema

CA :: n×3 Array{Float,2} Constantes de Antoine para los n componentes

Λ :: n×n Array{Float,2} Parametros de interacción entre los componentes en grados kelvin

xx :: StepRange ó Vector Son los puntos en el liquido donde se buscara el equilibrio con el vapor

uni Son las unidades en las que se desea el resultado, por defecto esta en grados kelvin

Salida:

Regresa las fracciones de liquido x, las fracciones del vapor y y la temperatura T en grados kelvin (m es la longitud de xx).

x :: n×m Array{Float64,2} liquido

y :: n×m Array{Float64,2} vapor

T :: m-element Array{Float64,1} temperatura

TermoQuimica.nrtl.pxy — Function

pxy(n,T,CA,Δg,α,xx;uni=u"Torr")

Los campos de entrada son:

n :: Int Es el numero de componentes

T :: Float Es la temperatura del sistema

CA :: n×3 Array{Float,2} Constantes de Antoine para los n componentes

Δg :: n×n Array{Float,2} Parametros de interacción entre los componentes en grados kelvin

α :: n×n Array{Float,2} Parametro de no aleatoriedad

xx :: StepRange ó Vector Son los puntos en el liquido donde se buscara el equilibrio con el vapor

uni Son las unidades en las que se desea el resultado, por defecto esta en Torr

Salida:

Regresa las fracciones de liquido x, las fracciones del vapor y y la presión P en Torr (m es la longitud de xx).

x :: n×m Array{Float64,2} liquido

y :: n×m Array{Float64,2} vapor

P :: m-element Array{Float64,1} presión

TermoQuimica.nrtl.txy — Function

txy(n,Pobj,CA,Δg,α,xx;uni=u"K")

Los campos de entrada son:

n :: Int Es el numero de componentes

Pobj :: Float Es la presión del sistema

CA :: n×3 Array{Float,2} Constantes de Antoine para los n componentes

Δg :: n×n Array{Float,2} Parametros de interacción entre los componentes en grados kelvin

α :: n×n Array{Float,2} Parametro de no aleatoriedad

xx :: StepRange ó Vector Son los puntos en el liquido donde se buscara el equilibrio con el vapor

uni Son las unidades en las que se desea el resultado, por defecto esta en grados kelvin

Salida:

Regresa las fracciones de liquido x, las fracciones del vapor y y la temperatura T en grados kelvin (m es la longitud de xx).

x :: n×m Array{Float64,2} liquido

y :: n×m Array{Float64,2} vapor

T :: m-element Array{Float64,1} temperatura

TermoQuimica.wilson.pxy — Function

pxy(n,T,CA,λ,v,xx;uni=u"Torr")

Los campos de entrada son:

n :: Int Es el numero de componentes

T :: Float Es la temperatura del sistema

CA :: n×3 Array{Float,2} Constantes de Antoine para los n componentes

λ :: n×n Array{Float,2} Parametros de interacción entre los componentes en grados kelvin

v :: n-element Array{Float,1} Volumen molar

xx :: StepRange ó Vector Son los puntos en el liquido donde se buscara el equilibrio con el vapor

uni Son las unidades en las que se desea el resultado, por defecto esta en Torr

Salida:

Regresa las fracciones de liquido x, las fracciones del vapor y y la presión P en Torr (m es la longitud de xx).

x :: n×m Array{Float64,2} liquido

y :: n×m Array{Float64,2} vapor

P :: m-element Array{Float64,1} presión

TermoQuimica.wilson.txy — Function

txy(n,Pobj,CA,λ,v,xx;uni=u"K")

Los campos de entrada son:

n :: Int Es el numero de componentes

Pobj :: Float Es la presión del sistema

CA :: n×3 Array{Float,2} Constantes de Antoine para los n componentes

λ :: n×n Array{Float,2} Parametros de interacción entre los componentes en grados kelvin

v :: n-element Array{Float,1} Volumen molar

xx :: StepRange ó Vector Son los puntos en el liquido donde se buscara el equilibrio con el vapor

uni Son las unidades en las que se desea el resultado, por defecto esta en grados kelvin

Salida:

Regresa las fracciones de liquido x, las fracciones del vapor y y la temperatura T en grados kelvin (m es la longitud de xx).

x :: n×m Array{Float64,2} liquido

y :: n×m Array{Float64,2} vapor

T :: m-element Array{Float64,1} temperatura

TermoQuimica.wilsonmod.pxy — Function

pxy(n,T,CA,a,b,c,v,xx;uni=u"Torr")

Los campos de entrada son:

n :: Int Es el numero de componentes

T :: Float Es la temperatura del sistema

CA :: n×3 Array{Float,2} Constantes de Antoine para los n componentes

a :: n×n Array{Float,2} Parametros de interacción a en $\frac{cal}{mol}$

b :: n×n Array{Float,2} Parametros de interacción b en $\frac{cal}{mol K}$

c :: n×n Array{Float,2} Parametros de interacción c en $\frac{cal}{mol K^2}$

v :: n-element Array{Float,1} Volumen molar

xx :: StepRange ó Vector Son los puntos en el liquido donde se buscara el equilibrio con el vapor

uni Son las unidades en las que se desea el resultado, por defecto esta en Torr

Salida:

Regresa las fracciones de liquido x, las fracciones del vapor y y la presión P en Torr (m es la longitud de xx).

x :: n×m Array{Float64,2} liquido

y :: n×m Array{Float64,2} vapor

P :: m-element Array{Float64,1} presión

TermoQuimica.wilsonmod.txy — Function

txy(n,Pobj,CA,a,b,c,v,xx;uni=u"K")

Los campos de entrada son:

n :: Int Es el numero de componentes

Pobj :: Float Es la presión del sistema

CA :: n×3 Array{Float,2} Constantes de Antoine para los n componentes

a :: n×n Array{Float,2} Parametros de interacción a en $\frac{cal}{mol}$

b :: n×n Array{Float,2} Parametros de interacción b en $\frac{cal}{mol K}$

c :: n×n Array{Float,2} Parametros de interacción c en $\frac{cal}{mol K^2}$

v :: n-element Array{Float,1} Volumen molar

xx :: StepRange ó Vector Son los puntos en el liquido donde se buscara el equilibrio con el vapor

uni Son las unidades en las que se desea el resultado, por defecto esta en grados kelvin

Salida:

Regresa las fracciones de liquido x, las fracciones del vapor y y la temperatura T en grados kelvin (m es la longitud de xx).

x :: n×m Array{Float64,2} liquido

y :: n×m Array{Float64,2} vapor

T :: m-element Array{Float64,1} temperatura

TermoQuimica.uniquac.pxy — Function

pxy(n,T,CA,Δu,q,r,xx;uni=u"Torr")

Los campos de entrada son:

n :: Int Es el numero de componentes

T :: Float Es la temperatura del sistema

CA :: n×3 Array{Float,2} Constantes de Antoine para los n componentes

Δu :: n×n Array{Float,2} Parametros de interacción entre los componentes en grados kelvin

q :: n-element Array{Float64,1} área de Van der Waals

r :: n-element Array{Float64,1} volumen de Van der Waals

xx :: StepRange ó Vector Son los puntos en el liquido donde se buscara el equilibrio con el vapor

uni Son las unidades en las que se desea el resultado, por defecto esta en Torr

Salida:

Regresa las fracciones de liquido x, las fracciones del vapor y y la presión P en Torr (m es la longitud de xx).

x :: n×m Array{Float64,2} liquido

y :: n×m Array{Float64,2} vapor

P :: m-element Array{Float64,1} presión

TermoQuimica.uniquac.txy — Function

txy(n,Pobj,CA,Δu,q,r,xx;uni=u"K")

Los campos de entrada son:

n :: Int Es el numero de componentes

Pobj :: Float Es la presión del sistema

CA :: n×3 Array{Float,2} Constantes de Antoine para los n componentes

Δu :: n×n Array{Float,2} Parametros de interacción entre los componentes en grados kelvin

q :: n-element Array{Float64,1} área de Van der Waals

r :: n-element Array{Float64,1} volumen de Van der Waals

xx :: StepRange ó Vector Son los puntos en el liquido donde se buscara el equilibrio con el vapor

uni Son las unidades en las que se desea el resultado, por defecto esta en grados kelvin

Salida:

Regresa las fracciones de liquido x, las fracciones del vapor y y la temperatura T en grados kelvin (m es la longitud de xx).

x :: n×m Array{Float64,2} liquido

y :: n×m Array{Float64,2} vapor

T :: m-element Array{Float64,1} temperatura

TermoQuimica.unifac.pxy — Function

pxy(n,T,CA,com,xx;uni=u"Torr")

Los campos de entrada son:

n :: Int Es el numero de componentes

T :: Float Es la temperatura del sistema

CA :: n×3 Array{Float,2} Constantes de Antoine para los n componentes

com ::

xx :: StepRange ó Vector Son los puntos en el liquido donde se buscara el equilibrio con el vapor

uni Son las unidades en las que se desea el resultado, por defecto esta en Torr

Salida:

Regresa las fracciones de liquido x, las fracciones del vapor y y la presión P en Torr (m es la longitud de xx).

x :: n×m Array{Float64,2} liquido

y :: n×m Array{Float64,2} vapor

P :: m-element Array{Float64,1} presión

TermoQuimica.unifac.txy — Function

txy(n,Pobj,CA,com,xx;uni=u"K")

Los campos de entrada son:

n :: Int Es el numero de componentes

Pobj :: Float Es la presión del sistema

CA :: n×3 Array{Float,2} Constantes de Antoine para los n componentes

com ::

xx :: StepRange ó Vector Son los puntos en el liquido donde se buscara el equilibrio con el vapor

uni Son las unidades en las que se desea el resultado, por defecto esta en grados kelvin

Salida:

Regresa las fracciones de liquido x, las fracciones del vapor y y la temperatura T en grados kelvin (m es la longitud de xx).

x :: n×m Array{Float64,2} liquido

y :: n×m Array{Float64,2} vapor

T :: m-element Array{Float64,1} temperatura

TermoQuimica.unifacmod.pxy — Function

pxy(n,T,CA,com,xx;uni=u"Torr")

Los campos de entrada son:

n :: Int Es el numero de componentes

T :: Float Es la temperatura del sistema

CA :: n×3 Array{Float,2} Constantes de Antoine para los n componentes

com ::

xx :: StepRange ó Vector Son los puntos en el liquido donde se buscara el equilibrio con el vapor

uni Son las unidades en las que se desea el resultado, por defecto esta en Torr

Salida:

Regresa las fracciones de liquido x, las fracciones del vapor y y la presión P en Torr (m es la longitud de xx).

x :: n×m Array{Float64,2} liquido

y :: n×m Array{Float64,2} vapor

P :: m-element Array{Float64,1} presión

TermoQuimica.unifacmod.txy — Function

txy(n,Pobj,CA,com,xx;uni=u"K")

Los campos de entrada son:

n :: Int Es el numero de componentes

Pobj :: Float Es la presión del sistema

CA :: n×3 Array{Float,2} Constantes de Antoine para los n componentes

com ::

xx :: StepRange ó Vector Son los puntos en el liquido donde se buscara el equilibrio con el vapor

uni Son las unidades en las que se desea el resultado, por defecto esta en grados kelvin

Salida:

Regresa las fracciones de liquido x, las fracciones del vapor y y la temperatura T en grados kelvin (m es la longitud de xx).

x :: n×m Array{Float64,2} liquido

y :: n×m Array{Float64,2} vapor

T :: m-element Array{Float64,1} temperatura

TermoQuimica.EOS.Van_Der_Waals — Function

Van_Der_Waals(T,P,Tc,Pc)

Ecuación de la forma:

\[P=\frac{RT}{ν-b}-\frac{a}{ν^2}\]

Donde:

\[\begin{aligned} a&=\frac{27(RT_c)^2}{64P_c} & b&=\frac{RT_c}{8P_c}\\ \end{aligned}\]

\[\frac{U-U^{ig}}{RT}=\frac{a}{RTν}\]

\[\frac{G-G^{ig}}{RT}=z-1-\frac{a}{RTν}-\log_{10} (z-B)\]

Los campos de entrada son:

T :: Float Es la temperatura del sistema

P :: Float Es la presión del sistema

Tc :: Float Es la temperatura critica del sistema

Pc :: Float Es la presión critica del sistema

𝜔 :: Float Es el factor acéntrico

Salida:

Regresa un conjunto de propiedades termodinámicas.

ℍ :: Float Entalpía residual

𝕌 :: Float Energía interna residual

𝕊 :: Float Entropía residual

𝔾 :: Float Energía libre de Gibbs residual

𝔸 :: Float Energía libre de helmholtz residual

φ :: Float Coeficiente de fugacidad

f :: Float Fugacidad

z :: Float Factor de compresibilidad

ν :: Float Volumen molar

TermoQuimica.EOS.Redlich_Kwong — Function

Redlich_Kwong(T,P,Tc,Pc)

Ecuación de la forma:

\[P=\frac{RT}{ν-b}-\frac{a}{(T^{1.5} ν(ν+b))}\]

Donde:

\[\begin{aligned} a&=\frac{0.42748 R^2 T_c^{2.5}}{P_c} & b&=\frac{0.08664 RT_c}{P_c}\\ \end{aligned}\]

\[\frac{U-U^{ig}}{RT}=\frac{3a\log_{10} (1+z/B)}{2bR T^{1.5}}\]

\[\frac{G-G^{ig}}{RT}=z-1+\frac{a\log_{10} (1+z/B)}{bR T^{1.5}}-\log_{10} (z-B)\]

Los campos de entrada son:

T :: Float Es la temperatura del sistema

P :: Float Es la presión del sistema

Tc :: Float Es la temperatura critica del sistema

Pc :: Float Es la presión critica del sistema

𝜔 :: Float Es el factor acéntrico

Salida:

Regresa un conjunto de propiedades termodinámicas.

ℍ :: Float Entalpía residual

𝕌 :: Float Energía interna residual

𝕊 :: Float Entropía residual

𝔾 :: Float Energía libre de Gibbs residual

𝔸 :: Float Energía libre de helmholtz residual

φ :: Float Coeficiente de fugacidad

f :: Float Fugacidad

z :: Float Factor de compresibilidad

v :: Float Volumen molar

TermoQuimica.EOS.Soave_Redlich_Kwong — Function

Soave_Redlich_Kwong(T,P,Tc,Pc,𝜔)

Ecuación de la forma:

\[P=\frac{RT}{V-b}-\frac{aα}{V(V+b)}\]

Donde:

\[\begin{aligned} a&=\frac{0.42748 R^2 T_c^{2}}{P_c} & b&=\frac{0.08664 RT_c}{P_c} & α&=(1 + κ(1 -T_r^{0.5}))^2\\ \end{aligned}\]

\[κ=0.48+1.574w-0.176w^2\]

\[\frac{U-U^{ig}}{RT}=-\frac{(aα+a\sqrt{T_rα}κ)\log_{10} (1+B/z)}{bRT}\]

\[\frac{S-S^{ig}}{R}=\log_{10}(z-B)-\frac{a\sqrt{T_rα}κ\log_{10}(1+B/z)}{bRT}\]

Los campos de entrada son:

T :: Float Es la temperatura del sistema

P :: Float Es la presión del sistema

Tc :: Float Es la temperatura critica del sistema

Pc :: Float Es la presión critica del sistema

𝜔 :: Float Es el factor acéntrico

Salida:

Regresa un conjunto de propiedades termodinámicas.

ℍ :: Float Entalpía residual

𝕌 :: Float Energía interna residual

𝕊 :: Float Entropía residual

𝔾 :: Float Energía libre de Gibbs residual

𝔸 :: Float Energía libre de helmholtz residual

φ :: Float Coeficiente de fugacidad

f :: Float Fugacidad

z :: Float Factor de compresibilidad

v :: Float Volumen molar

TermoQuimica.EOS.Peng_Robinson — Function

Peng_Robinson(T,P,Tc,Pc,𝜔)

Ecuación de la forma:

\[P=\frac{RT}{V-b}-\frac{aα}{V^2 +2bV+b^2}\]

Donde:

\[\begin{aligned} a&=\frac{0.45724 R^2 T_c^{2}}{P_c} & b&=\frac{0.0778 RT_c}{P_c} & α&=(1 + κ(1 -T_r^{0.5}))^2\\ \end{aligned}\]

\[κ=0.37464 + 1.54226*w -0.26992*w^2\]

\[\frac{U-U^{ig}}{RT}=-(aα+a\sqrt{T_rα}κ)*(\coth^{-1}{\sqrt{2}}+\tanh^{-1}\frac{(-1+B/z)/\sqrt{2}}{bRT\sqrt{2}})\]

\[\frac{G-G^{ig}}{R}=z-1-\log(z-B)-aα*(\coth^{-1}{\sqrt{2}}+\tanh^{-1}\frac{(B-z)/z\sqrt{2}}{bRT\sqrt{2}})\]

Los campos de entrada son:

T :: Float Es la temperatura del sistema

P :: Float Es la presión del sistema

Tc :: Float Es la temperatura critica del sistema

Pc :: Float Es la presión critica del sistema

𝜔 :: Float Es el factor acéntrico

Salida:

Regresa un conjunto de propiedades termodinámicas.

ℍ :: Float Entalpía residual

𝕌 :: Float Energía interna residual

𝕊 :: Float Entropía residual

𝔾 :: Float Energía libre de Gibbs residual

𝔸 :: Float Energía libre de helmholtz residual

φ :: Float Coeficiente de fugacidad

f :: Float Fugacidad

z :: Float Factor de compresibilidad

v :: Float Volumen molar

Missing docstring.

Missing docstring for accesorios.entrada_recta. Check Documenter's build log for details.